requirements.txt
hcloud>=2.0.0
PyYAML>=6.0
requests>=2.31.0
